Cosmetic change: renamed namespaces for examples
